requirements.txt
sentence-transformers
faiss-cpu
numpy
scikit-learn
pandas
matplotlib